requirements.txt
streamlit==1.39.0
pandas==2.2.2
plotly==5.24.1
folium==0.17.0
streamlit-folium==0.20.0
Pillow==10.4.0
numpy==1.26.4